Week-9/Day-5/requirements.txt
pyautogen>=0.2.0
streamlit>=1.30.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
groq>=0.9.0
torchvision
